Move nias_cpp support classes to separate files
