+++ src/nias/bindings/nias_cpp/vector.py
@@ -1,9 +1,6 @@
-import numpy as np
-from nias.interfaces import InnerProduct, VectorArray, VectorSpace
-from numpy.typing import NDArray
+from nias.interfaces import VectorArray, VectorSpace
 
 from nias.bindings.nias_cpp.interfaces import ListVectorArray, ListVectorArrayImpl, ListVectorSpace, Vector
-
 
 class NiasCppVector(Vector):
     def __init__(self, impl):
@@ -76,15 +73,3 @@ class NiasCppListVectorSpace(ListVectorSpace):
             vecs = [vecs]
         return ListVectorArray(impl=ListVectorArrayImpl(vecs, len(vecs[0])))
 
-
-class NiasCppInnerProduct(InnerProduct):
-    def apply(self, left: VectorArray, right: VectorArray, pairwise: bool = False) -> NDArray:
-        if pairwise:
-            ret = [left_vec.dot(right_vec) for left_vec, right_vec in zip(left.vectors, right.vectors)]
-            return np.array([ret])
-        else:
-            ret = np.zeros(shape=(len(left), len(right)))
-            for i, left_vec in enumerate(left.vectors):
-                for j, right_vec in enumerate(right.vectors):
-                    ret[i, j] = left_vec.dot(right_vec)
-            return ret
